% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_spatialdata_tutorial.R
\name{run_spatialdata_tutorial}
\alias{run_spatialdata_tutorial}
\title{Run a Building Spatial Data interactive tutorial}
\usage{
run_spatialdata_tutorial(name, ...)
}
\arguments{
\item{name}{Tutorial name.}

\item{...}{Additional parameters passed to \code{\link[learnr:run_tutorial]{learnr::run_tutorial()}}}
}
\value{
Launch a learnr exercise.
}
\description{
A wrapper for \code{\link[learnr:run_tutorial]{learnr::run_tutorial()}} with validation for the tutorial name.
}
\examples{
\dontrun{
if(interactive()){
 run_spatialdata_tutorial(name = "summarise_sf")
 }
}
}
\seealso{
\code{\link[learnr]{run_tutorial}}
}
